일별공유이력 추가 구매시 인플루언서 포인트 지급 및 이력생성 구매취소시 인플루언서 포인트회수 및 이력생성 (테스트 강도높게 할 부분)

+++ src/main/java/com/usit/app/spring/security/domain/SignedMember.java
@@ -64,7 +64,7 @@ public class SignedMember extends UserDetailsWrapper{
         return this.vo.getMemberId();
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.vo.setMemberId(userId);
     }
