+++ src/main/java/com/usit/domain/ProductOption.java
@@ -69,14 +69,14 @@ public class ProductOption {
 	private LocalDateTime modDate;
 
 	@Column(name="mod_id")
-	private Long modId;
+	private Integer modId;
 
 	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
 	@Column(name="reg_date")
 	private LocalDateTime regDate;
 
 	@Column(name="reg_id")
-	private Long regId;
+	private Integer regId;
 	
 	
 	@PrePersist
