+++ src/main/java/com/usit/repository/OrderRepository.java
@@ -13,11 +13,11 @@ import com.usit.domain.UsitOrder;
 @Repository
 public interface OrderRepository extends JpaRepository<UsitOrder, Integer>{
 
-    public Page<UsitOrder> findByMemberIdAndOrderStatusCdNotAndUseYn(Long memberId, Pageable pageable,String orderStatusCd,String useYn);
-    public Page<UsitOrder> findByMemberId(Long memberId, Pageable pageable);
-    public Page<UsitOrder> findByMemberIdAndOrderStatusCdNot(Long memberId, String orderStatusCd, Pageable pageable);
+    public Page<UsitOrder> findByMemberIdAndOrderStatusCdNotAndUseYn(int memberId, Pageable pageable,String orderStatusCd,String useYn);
+    public Page<UsitOrder> findByMemberId(int memberId, Pageable pageable);
+    public Page<UsitOrder> findByMemberIdAndOrderStatusCdNot(int memberId, String orderStatusCd, Pageable pageable);
     public Page<UsitOrder> findAllByOrderStatusCdNot(Pageable pageable,String orderStatusCd);
-    public List<UsitOrder> findAllByMemberId(Long memberId);
+    public List<UsitOrder> findAllByMemberId(int memberId);
  
 //    public Page<My23Order> findAllByAndDeliverStatusCdInOrderStatusCdNot(Pageable pageable, List<String> deliveryStatusCds, String orderStatusCd);
     public UsitOrder findFirstByMerchantUid(String merchantUid);
