+++ src/main/java/com/usit/repository/SellMemberRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.usit.domain.SellMember;
 
 @Repository
-public interface SellMemberRepository extends JpaRepository<SellMember, Long>{
+public interface SellMemberRepository extends JpaRepository<SellMember, Integer>{
 
 	public SellMember findBySellMemberUid(String uid);
 	public SellMember findByEmail(String email);
